+++ moll/picking/tests/online_picker_test.py
@@ -6,9 +6,9 @@ import jax.numpy as jnp
 import pytest
 from sklearn import datasets  # type: ignore
 
-from ..metrics import euclidean, one_minus_tanimoto
-from ..utils.utils import generate_points, random_grid_points
-from .online_picker import OnlineDiversityPicker
+from ...metrics import euclidean, one_minus_tanimoto
+from ...utils.utils import generate_points, random_grid_points
+from ..online_picker import OnlineDiversityPicker
 
 RANDOM_SEED = 42
